[Decomp][ObjectMgr] Cohesion split: extract quest loaders to ObjectMgrQuests.cpp
